+++ app/[locale]/account/profile/__tests__/page.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {act, fireEvent, render, screen, within} from '@testing-library/react';
+import {act, fireEvent, render, screen, within} from '@/utils/test-utils';
 import {axe, toHaveNoViolations} from 'jest-axe';
 import ProfilePage from '../page';
 import {
@@ -19,7 +19,6 @@ jest.mock('@/utils/clientLogger', () => ({
   default: jest.fn()
 }))
 
-
 // Mock the GraphQL hooks
 jest.mock('@/generated/graphql', () => ({
   useMeQuery: jest.fn(),
@@ -37,6 +36,22 @@ jest.mock('@/components/UpdateEmailAddress', () => ({
 jest.mock('@/components/Form/TypeAheadWithOther', () => ({
   __esModule: true,
   default: () => <div data-testid="type-ahead">Mocked TypeAheadWithOther Component</div>,
+}));
+
+jest.mock('@/i18n/routing', () => ({
+  usePathname: jest.fn(() => '/about'),
+}));
+
+jest.mock('@/components/PageHeader');
+
+
+// Mock useFormatter and useTranslations from next-intl
+jest.mock('next-intl', () => ({
+  useFormatter: jest.fn(() => ({
+    dateTime: jest.fn(() => '01-01-2023'),
+  })),
+  useTranslations: jest.fn(() => jest.fn((key) => key)), // Mock `useTranslations`,
+  useLocale: jest.fn(() => 'en-US'), // Return a default locale
 }));
 
 const mockUserData = {
@@ -74,26 +89,10 @@ describe('ProfilePage', () => {
   it('should render profile page with user data', async () => {
     render(<ProfilePage />);
 
-    await act(async () => {
-      const element = screen.getByText('Update profile', { selector: '.pageheader-title' });
-      expect(element).toBeInTheDocument();
-    });
-
     // Check that user data is rendered
     expect(screen.getByText('John')).toBeInTheDocument();
     expect(screen.getByText('Doe')).toBeInTheDocument();
     expect(screen.getByText('Test Institution')).toBeInTheDocument();
-  });
-
-  it('sets the document title correctly', () => {
-    render(<ProfilePage />);
-    expect(document.title).toBe('Update profile | DMPTool');
-  });
-
-  it('should scroll to the top of the page on render', () => {
-    window.scrollTo = jest.fn();
-    render(<ProfilePage />);
-    expect(window.scrollTo).toHaveBeenCalledWith(0, 0);
   });
 
   it('should show form fields when Edit button is clicked', async () => {