+++ app/[locale]/account/profile/page.tsx
@@ -2,7 +2,10 @@
 
 import React, {useEffect, useState} from 'react';
 import {ApolloError} from '@apollo/client';
+import {useRouter} from 'next/navigation';
 import Link from 'next/link';
+import {useLocale, useTranslations} from 'next-intl';
+import {usePathname} from '@/i18n/routing';
 import {
   Breadcrumb,
   Breadcrumbs,
@@ -42,9 +45,14 @@ import {
 
 // Utils and other
 import logECS from '@/utils/clientLogger';
+import {refreshAuthTokens} from "@/utils/authHelper";
 import styles from './profile.module.scss';
 
 const ProfilePage: React.FC = () => {
+  const t = useTranslations('UserProfile');
+  const pathname = usePathname();
+  const currentLocale = useLocale();
+  const router = useRouter();
   const [otherField, setOtherField] = useState(false);
   // We need to save the original data for when users cancel their form updates
   const [originalData, setOriginalData] = useState<ProfileDataInterface>();
@@ -72,6 +80,13 @@ const ProfilePage: React.FC = () => {
   });
   const [emailAddresses, setEmailAddresses] = useState<EmailInterface[]>([]);
   const [languages, setLanguages] = useState<LanguageInterface[]>([]);
+
+  const switchLanguage = (newLocale: string) => {
+    if (newLocale !== currentLocale) {
+      const newPath = `/${newLocale}${pathname}`;
+      router.push(newPath);
+    }
+  };
 
   // Initialize user profile mutation
   const [updateUserProfileMutation, { loading: updateUserProfileLoading }] = useUpdateUserProfileMutation();
@@ -137,6 +152,10 @@ const ProfilePage: React.FC = () => {
       const response = await profileUpdateMutation();
       if (response) {
         setIsEditing(false);
+        // Refresh token to include preferred language in token
+        await refreshAuthTokens();
+        // Update pathname to match the selected language so user can see page in selected language
+        switchLanguage(formData.languageId);
       }
     } catch (error) {
       if (error instanceof ApolloError) {
@@ -259,54 +278,55 @@ const ProfilePage: React.FC = () => {
   }, [languageData]);
 
   useEffect(() => {
-    //When data from backend changes, set formData and originalData
-    if (data && data.me) {
-      // Get name of selected language
-      const language = languages.find(lang => lang.id === data?.me?.languageId) || {
-        id: '',
-        name: '',
-        isDefault: false
-      };
+    const updateFormData = async () => {
+      if (data && data.me && languageData) {
+        // Get local languages that are not null (for Type guard)
+        const loadedLanguages = (languageData.languages || []).filter((language) => language !== null);
 
-      // Set email list so we can pass to the UpdateEmailAddresses component
-      if (data?.me?.emails) {
-        const validEmails = data.me.emails
-          // Filter out nulls and map to Email type
-          .filter((email): email is NonNullable<typeof email> => email !== null)
-          .map(email => ({
-            id: email.id ?? undefined,
-            email: email.email,
-            isPrimary: email.isPrimary,
-            isConfirmed: email.isConfirmed
-          }));
+        // Find the selected language or fallback
+        const selectedLanguage = loadedLanguages.find((lang) => lang.id === data?.me?.languageId) || {
+          id: '',
+          name: '',
+          isDefault: false,
+        };
 
-        setEmailAddresses(validEmails);
-      } else {
-        setEmailAddresses([]); // Reset to empty array if no email data
+        // Set user email list from data
+        if (data.me.emails) {
+          const validEmails = data.me.emails
+            .filter((email): email is NonNullable<typeof email> => email !== null)
+            .map((email) => ({
+              id: email.id ?? undefined,
+              email: email.email,
+              isPrimary: email.isPrimary,
+              isConfirmed: email.isConfirmed,
+            }));
+          setEmailAddresses(validEmails);
+        } else {
+          setEmailAddresses([]);
+        }
+
+        // Update originalData and formData
+        const newOriginalData = {
+          firstName: data.me.givenName ?? '',
+          lastName: data.me.surName ?? '',
+          affiliationName: data.me.affiliation?.name ?? '',
+          affiliationId: data.me.affiliation?.uri ?? '',
+          otherAffiliationName: '',
+          languageId: data.me.languageId,
+          languageName: selectedLanguage.name,
+        };
+
+        setOriginalData(newOriginalData);
+
+        setFormData((prev) => ({
+          ...prev,
+          ...newOriginalData,
+        }));
       }
+    };
 
-      setOriginalData({
-        firstName: data.me.givenName ?? '',
-        lastName: data.me.surName ?? '',
-        affiliationName: data.me.affiliation?.name ?? '',
-        affiliationId: data.me.affiliation?.uri ?? '',
-        otherAffiliationName: '',
-        languageId: data.me.languageId,
-        languageName: language.name
-      });
-
-      setFormData({
-        ...formData,
-        firstName: data.me.givenName ?? '',
-        lastName: data.me.surName ?? '',
-        affiliationName: data.me.affiliation?.name ?? '',
-        affiliationId: data.me.affiliation?.uri ?? '',
-        otherAffiliationName: '',
-        languageId: data.me.languageId,
-        languageName: language.name
-      })
-    }
-  }, [data])
+    updateFormData();
+  }, [data, languageData]);
 
   // Update form data
   const handleUpdate = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -340,12 +360,12 @@ const ProfilePage: React.FC = () => {
   return (
     <>
       <PageHeader
-        title="Update profile"
+        title={t('headingUpdateProfile')}
         showBackButton={true}
         breadcrumbs={
           <Breadcrumbs>
-            <Breadcrumb><Link href="/">Account</Link></Breadcrumb>
-            <Breadcrumb><Link href="/account/profile">Update Profile</Link></Breadcrumb>
+            <Breadcrumb><Link href="/">{t('breadcrumbHome')}</Link></Breadcrumb>
+            <Breadcrumb><Link href="/account/profile">{t('headingUpdateProfile')}</Link></Breadcrumb>
           </Breadcrumbs>
         }
         className="page-template-list"
@@ -354,7 +374,7 @@ const ProfilePage: React.FC = () => {
         <div className={styles.mainContent}>
           <LayoutWithPanel>
             <ContentContainer className={styles.layoutContentContainer}>
-              <h2>Your Profile</h2>
+              <h2>{t('yourProfile')}</h2>
               <div className="sectionContainer">
                 <div className={`sectionContent ${styles.section}`}>
                   <Form onSubmit={handleProfileSubmit}>
@@ -379,7 +399,7 @@ const ProfilePage: React.FC = () => {
                         />
                       ) : (
                         <Text slot="firstName" className={styles.readOnlyField}>
-                          <div className="field-label">First name</div>
+                          <div className="field-label">{t('firstName')}</div>
                           <p>{formData.firstName}</p>
                         </Text>
                       )}
@@ -397,7 +417,7 @@ const ProfilePage: React.FC = () => {
                         />
                       ) : (
                         <Text slot="lastName" className={styles.readOnlyField}>
-                          <div className="field-label">Last name</div>
+                          <div className="field-label">{t('lastName')}</div>
                           <p>{formData.lastName}</p>
                         </Text>
                       )}
@@ -414,7 +434,7 @@ const ProfilePage: React.FC = () => {
                             setOtherField={setOtherField}
                             required={true}
                             error={fieldErrors.affiliationName}
-                            helpText="Search for your institution"
+                            helpText={t('helpTextSearchForInstitution')}
                             updateFormData={updateAffiliationFormData}
                             value={formData.affiliationName}
                           />
@@ -436,7 +456,7 @@ const ProfilePage: React.FC = () => {
                         </>
                       ) : (
                         <Text slot="institution" className={styles.readOnlyField}>
-                          <div className="field-label">Institution</div>
+                          <div className="field-label">{t('institution')}</div>
                           <p>{formData.affiliationName}</p>
                         </Text>
                       )}
@@ -450,7 +470,7 @@ const ProfilePage: React.FC = () => {
                           name="institution"
                           items={languages}
                           errorMessage="A selection is required"
-                          helpMessage="Select your preferred language"
+                          helpMessage={t('helpTextSelectYourLanguage')}
                           onSelectionChange={selected => setFormData({ ...formData, languageId: selected as string })}
                           selectedKey={formData.languageId.trim()}
                         >
@@ -463,19 +483,19 @@ const ProfilePage: React.FC = () => {
                         </FormSelect>
                       ) : (
                         <Text slot="language" className={styles.readOnlyField}>
-                          <div className="field-label">Language</div>
+                          <div className="field-label">{t('language')}</div>
                           <p>{formData.languageName}</p>
                         </Text>
                       )}
                     </div>
                     {isEditing ? (
                       <div className={styles.btnContainer}>
-                        <Button className="secondary" onPress={cancelEdit}>Cancel</Button>
-                        <Button type="submit" isDisabled={updateUserProfileLoading} className={styles.btn}>{updateUserProfileLoading ? 'Updating' : 'Update'}</Button>
+                        <Button className="secondary" onPress={cancelEdit}>{t('btnCancel')}</Button>
+                        <Button type="submit" isDisabled={updateUserProfileLoading} className={styles.btn}>{updateUserProfileLoading ? t('btnUpdating') : t('btnUpdate')}</Button>
                       </div>
                     ) : (
                       <div className={styles.btnContainer}>
-                        <Button type="submit" onPress={handleEdit} className={styles.btn}>Edit</Button>
+                        <Button type="submit" onPress={handleEdit} className={styles.btn}>{t('btnEdit')}</Button>
                       </div>
                     )}
                   </Form>
@@ -487,16 +507,12 @@ const ProfilePage: React.FC = () => {
               />
             </ContentContainer>
             <SidebarPanel className={styles.layoutSidebarPanel}>
-
-
-              <h2>Related actions</h2>
+              <h2>{t('headingRelatedActions')}</h2>
               <ul className={styles.relatedItems}>
-                <li><Link href="/account/update-password">Update password</Link></li>
-                <li><Link href="/account/connections">Update connections</Link></li>
-                <li><Link href="/account/notifications">Manage notifications</Link></li>
+                <li><Link href="/account/update-password">{t('linkUpdatePassword')}</Link></li>
+                <li><Link href="/account/connections">{t('linkUpdateConnections')}</Link></li>
+                <li><Link href="/account/notifications">{t('linkManageNotifications')}</Link></li>
               </ul>
-
-
             </SidebarPanel>
           </LayoutWithPanel>
         </div>