+++ app/[locale]/page.tsx
@@ -1,8 +1,13 @@
+import {ContentContainer, LayoutContainer} from '@/components/Container';
 
 export default async function Home() {
 
 
   return (
-    <h1>Home Page</h1>
+    <LayoutContainer>
+      <ContentContainer>
+        <h1>Home Page</h1>
+      </ContentContainer>
+    </LayoutContainer>
   )
 }