+++ i18n/request.ts
@@ -11,14 +11,18 @@ export default getRequestConfig(async ({requestLocale}) => {
     locale = routing.defaultLocale;
   }
   // Load multiple translation files
-  const mainMessages = (await import(`@/messages/${locale}/en-US.json`)).default;
+  const mainMessages = (await import(`@/messages/${locale}/global.json`)).default;
+  const templateMessages = (await import(`@/messages/${locale}/templateBuilder.json`)).default;
   const planBuilderProjectOverviewMessages = (await import(`@/messages/${locale}/planBuilderProjectOverview.json`)).default;
   const planBuilderPlanOverviewMessages = (await import(`@/messages/${locale}/planBuilderPlanOverview.json`)).default;
   const errorMessages = (await import(`@/messages/${locale}/errors.json`)).default;
+  const messaging = (await import(`@/messages/${locale}/messaging.json`)).default;
 
   // Merge all message objects
   const messages = {
     ...mainMessages,
+    ...templateMessages,
+    ...messaging,
     ...planBuilderProjectOverviewMessages,
     ...planBuilderPlanOverviewMessages,
     errors: errorMessages,